usb: split CDC and UART implementations into separate translation units
